Add parameter to enable repo popula

+++ lib/builder.sh
@@ -420,18 +420,20 @@ function kernelPost {
     kernelBuild "`basename ${TMP_I}`"
   done
   
-  printStatus "armStrapPost" "Publishing kernel installer script"
-  checkDirectory "${ARMSTRAP_ABUILDER_KERNEL}"
-  for TMP_I in ${ARMSTRAP_PKG}/*.sh; do
-   mv -v ${TMP_I} ${ARMSTRAP_ABUILDER_KERNEL}/
-  done
+  if [ ! -z ${ARMSTRAP_ABUILDER_REPO_ENABLE} ]; then
+    printStatus "armStrapPost" "Publishing kernel installer script"
+    checkDirectory "${ARMSTRAP_ABUILDER_KERNEL}"
+    for TMP_I in ${ARMSTRAP_PKG}/*.sh; do
+      mv -v ${TMP_I} ${ARMSTRAP_ABUILDER_KERNEL}/
+    done
   
-  printStatus "armStrapPost" "Publishing kernels"
-  for TMP_I in ${ARMSTRAP_PKG}/*.deb; do
-    repoPost "${TMP_I}"
-    rm -f ${TMP_I}
-  done
-  indexPost
+    printStatus "armStrapPost" "Publishing kernels"
+    for TMP_I in ${ARMSTRAP_PKG}/*.deb; do
+      repoPost "${TMP_I}"
+      rm -f ${TMP_I}
+    done
+    indexPost
+  fi
 }
 
 function loaderPost {
@@ -442,8 +444,10 @@ function loaderPost {
   for TMP_I in ${ARMSTRAP_BOOTLOADERS}/*; do
     for TMP_J in ${TMP_I}/*; do
       bootBuilder "`basename ${TMP_I}`" "`basename ${TMP_J}`"
-      printStatus "armStrapPost" "Publishing bootloaders"
-      mv -v ${ARMSTRAP_PKG}/`basename ${TMP_J}`-`basename ${TMP_I}`${ARMSTRAP_TAR_EXTENSION} ${ARMSTRAP_ABUILDER_LOADER}/
+      if [ ! -z ${ARMSTRAP_ABUILDER_REPO_ENABLE} ]; then
+        printStatus "armStrapPost" "Publishing bootloaders"
+        mv -v ${ARMSTRAP_PKG}/`basename ${TMP_J}`-`basename ${TMP_I}`${ARMSTRAP_TAR_EXTENSION} ${ARMSTRAP_ABUILDER_LOADER}/
+      fi
     done
   done
   indexPost
@@ -457,8 +461,10 @@ function rootfsPost {
   for TMP_I in ${ARMSTRAP_ROOTFS}/*; do
     for TMP_J in ${TMP_I}/*; do
       rootfsUpdater "`basename ${TMP_J}`" "`basename ${TMP_I}`"
-      printStatus "armStrapPost" "Publishing rootfs"
-      mv -v ${ARMSTRAP_PKG}/`basename ${TMP_I}`-*-`basename ${TMP_J}`${ARMSTRAP_TAR_EXTENSION} ${ARMSTRAP_ABUILDER_ROOTFS}/
+      if [ ! -z ${ARMSTRAP_ABUILDER_REPO_ENABLE} ]; then
+        printStatus "armStrapPost" "Publishing rootfs"
+        mv -v ${ARMSTRAP_PKG}/`basename ${TMP_I}`-*-`basename ${TMP_J}`${ARMSTRAP_TAR_EXTENSION} ${ARMSTRAP_ABUILDER_ROOTFS}/
+      fi
     done
   done
   indexPost
@@ -468,14 +474,16 @@ function indexPost {
   local TMP_I=""
   local TMP_J=""
   
-  printStatus "armStrapPost" "Making indexes"
-  rm -f ${ARMSTRAP_ABUILDER_ROOT}/.index.txt
-  touch ${ARMSTRAP_ABUILDER_ROOT}/.index.txt
-  for TMP_I in ${ARMSTRAP_ABUILDER_KERNEL} ${ARMSTRAP_ABUILDER_ROOTFS} ${ARMSTRAP_ABUILDER_LOADER}; do
-    for TMP_J in `find ${TMP_I} -type f | sort`; do
-      echo ${TMP_J/$ARMSTRAP_ABUILDER_ROOT/} >> ${ARMSTRAP_ABUILDER_ROOT}/.index.txt
+  if [ ! -z ${ARMSTRAP_ABUILDER_REPO_ENABLE} ]; then
+    printStatus "armStrapPost" "Making indexes"
+    rm -f ${ARMSTRAP_ABUILDER_ROOT}/.index.txt
+    touch ${ARMSTRAP_ABUILDER_ROOT}/.index.txt
+    for TMP_I in ${ARMSTRAP_ABUILDER_KERNEL} ${ARMSTRAP_ABUILDER_ROOTFS} ${ARMSTRAP_ABUILDER_LOADER}; do
+      for TMP_J in `find ${TMP_I} -type f | sort`; do
+        echo ${TMP_J/$ARMSTRAP_ABUILDER_ROOT/} >> ${ARMSTRAP_ABUILDER_ROOT}/.index.txt
+      done
     done
-  done
+  fi
 }
 
 function armStrapBuild {
